forseti-iracing-bridge/requirements.txt
pyirsdk>=1.3.5
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
